refactor(desktop): retire standalone plugins/skills view for Tool Access
- Route sidebar Plugins nav and openSkills through Settings > Tool Access
- Delete SkillsView, PluginsCatalogPage, SkillsCatalogPage, card grids, and
  plugin management workspace state (pluginManagementWorkspaceId/Mode)
- Move still-used dialogs (ImportDialog, Install/Detail dialogs, skillUtils)
  under src/ui/settings/toolAccess/
- Remove dead WorkspacesPage surface=all tab shell; relocate per-workspace
  settings toggle into the defaults Advanced actions collapsible
- Update tests: drop catalog page/plugin management suites, retarget dialog
  imports, remap legacy skills view cache assertions to settings
