+++ apps/desktop/src/app/store.helpers/runtimeState.ts
@@ -335,7 +335,6 @@ export function defaultWorkspaceRuntime(): WorkspaceRuntime {
     pluginMutationError: null,
     importItemsByKey: {},
     importPendingKeys: {},
-    pluginViewMode: "plugins",
     memories: [],
     memoriesLoading: false,
     advancedMemories: [],
