+++ apps/desktop/src/styles/platform/linux.css
@@ -224,10 +224,6 @@
   overflow: hidden;
 }
 
-:root[data-platform="linux"] .settings-shell__header-actions {
-  padding-right: calc(var(--platform-caption-button-reserve) + var(--platform-topbar-toolbar-gap));
-}
-
 :root[data-platform="linux"] .settings-shell__drag-zone {
   position: absolute;
   top: 0;
