+++ apps/desktop/src/ui/settings/toolAccess/PluginDetailDialog.tsx
@@ -1,19 +1,19 @@
 import { ExternalLinkIcon, PackageIcon } from "lucide-react";
 import { useMemo } from "react";
 
-import { useAppStore } from "../../app/store";
-import { Badge } from "../../components/ui/badge";
-import { Button } from "../../components/ui/button";
+import { useAppStore } from "../../../app/store";
+import { Badge } from "../../../components/ui/badge";
+import { Button } from "../../../components/ui/button";
 import {
   Dialog,
   DialogContent,
   DialogDescription,
   DialogHeader,
   DialogTitle,
-} from "../../components/ui/dialog";
-import { confirmAction, revealPath } from "../../lib/desktopCommands";
-import { isInstalledPluginCatalogEntry } from "../../lib/wsProtocol";
-import { actionPending } from "../skills/utils";
+} from "../../../components/ui/dialog";
+import { confirmAction, revealPath } from "../../../lib/desktopCommands";
+import { isInstalledPluginCatalogEntry } from "../../../lib/wsProtocol";
+import { actionPending } from "./skillUtils";
 
 function pluginScopeLabel(scope: "workspace" | "user"): string {
   return scope === "workspace" ? "Workspace" : "Library";
