--- apps/desktop/src/ui/skills/SkillsCatalogPage.tsx
@@ -1,185 +1,0 @@
-import { RefreshCwIcon } from "lucide-react";
-import { useMemo } from "react";
-import { useAppStore } from "../../app/store";
-import { Badge } from "../../components/ui/badge";
-import { Button } from "../../components/ui/button";
-import type { MarketplaceSkillCatalogEntry } from "../../lib/wsProtocol";
-import { InlineErrorBoundary } from "../CrashReportingErrorBoundary";
-import { ImportDialog } from "../import/ImportDialog";
-import { AvailableSkillCardGrid } from "./AvailableSkillCardGrid";
-import { InstallationCardGrid } from "./InstallationCardGrid";
-import { InstallSkillDialog } from "./InstallSkillDialog";
-import { SkillDetailDialog } from "./SkillDetailDialog";
-
-export function SkillsCatalogPage({
-  workspaceId,
-  managementScope = "workspace",
-  searchQuery,
-}: {
-  workspaceId: string;
-  managementScope?: "workspace" | "global";
-  searchQuery: string;
-}) {
-  const wsRtById = useAppStore((s) => s.workspaceRuntimeById);
-  const selectSkillInstallation = useAppStore((s) => s.selectSkillInstallation);
-  const refreshSkillsCatalog = useAppStore((s) => s.refreshSkillsCatalog);
-  const installSkills = useAppStore((s) => s.installSkills);
-
-  const rt = wsRtById[workspaceId];
-  const catalog = rt?.skillsCatalog ?? null;
-  const skillCatalogLoading = rt?.skillCatalogLoading ?? false;
-  const skillCatalogError = rt?.skillCatalogError ?? null;
-  const showLoadingState = skillCatalogLoading && catalog === null;
-
-  const installations = useMemo(() => {
-    let items = [...(catalog?.installations ?? [])];
-    if (managementScope === "global") {
-      items = items.filter(
-        (installation) => installation.scope === "global" || installation.scope === "user",
-      );
-    }
-
-    if (searchQuery.trim()) {
-      const q = searchQuery.toLowerCase();
-      items = items.filter(
-        (i) =>
-          i.name.toLowerCase().includes(q) ||
-          i.interface?.displayName?.toLowerCase().includes(q) ||
-          i.description?.toLowerCase().includes(q) ||
-          i.interface?.shortDescription?.toLowerCase().includes(q),
-      );
-    }
-
-    return items.sort((left, right) =>
-      `${left.name}:${left.scope}:${left.installationId}`.localeCompare(
-        `${right.name}:${right.scope}:${right.installationId}`,
-      ),
-    );
-  }, [catalog, managementScope, searchQuery]);
-
-  const effectiveSkills = useMemo(() => {
-    return installations.filter((i) => i.state === "effective");
-  }, [installations]);
-
-  const otherSkills = useMemo(() => {
-    return installations.filter((i) => i.state !== "effective");
-  }, [installations]);
-
-  const availableSkills = useMemo(() => {
-    let items = [...(catalog?.availableSkills ?? [])];
-    if (searchQuery.trim()) {
-      const q = searchQuery.toLowerCase();
-      items = items.filter(
-        (s) =>
-          s.name.toLowerCase().includes(q) ||
-          s.displayName.toLowerCase().includes(q) ||
-          s.description.toLowerCase().includes(q) ||
-          s.category.toLowerCase().includes(q),
-      );
-    }
-    return items.sort((left, right) => left.displayName.localeCompare(right.displayName));
-  }, [catalog, searchQuery]);
-
-  const installingSkill = Object.keys(rt?.skillMutationPendingKeys ?? {}).some((key) =>
-    key.startsWith("install:"),
-  );
-
-  const handleInstallAvailableSkill = (skill: MarketplaceSkillCatalogEntry) => {
-    // Marketplace skills are user-level offerings; install them globally.
-    void installSkills(skill.installSource, "global").catch(() => {});
-  };
-
-  return (
-    <div className="app-skills-view h-full min-h-0 overflow-y-auto px-6 py-4">
-      <div className="mx-auto max-w-6xl">
-        <div className="mb-4 flex items-center justify-end gap-2">
-          <Button
-            variant="ghost"
-            size="sm"
-            className="h-8 px-2 text-muted-foreground hover:text-foreground"
-            onClick={() => void refreshSkillsCatalog()}
-          >
-            <RefreshCwIcon className="mr-1.5 h-4 w-4" />
-            Refresh
-          </Button>
-          <ImportDialog workspaceId={workspaceId} kind="skill" />
-          <InstallSkillDialog workspaceId={workspaceId} />
-        </div>
-
-        <InlineErrorBoundary label="The skills catalog couldn't be loaded.">
-          <div className="space-y-6">
-            {effectiveSkills.length > 0 && (
-              <section>
-                <h2 className="text-lg font-semibold mb-4">Installed</h2>
-                <InstallationCardGrid
-                  installations={effectiveSkills}
-                  onSelect={(id) => void selectSkillInstallation(id)}
-                />
-              </section>
-            )}
-
-            {otherSkills.length > 0 && (
-              <section>
-                <h2 className="mb-4 text-lg font-semibold text-muted-foreground">
-                  Other Installations
-                </h2>
-                <InstallationCardGrid
-                  installations={otherSkills}
-                  onSelect={(id) => void selectSkillInstallation(id)}
-                />
-              </section>
-            )}
-
-            {availableSkills.length > 0 && (
-              <section>
-                <h2 className="mb-4 text-lg font-semibold">Available</h2>
-                <AvailableSkillCardGrid
-                  skills={availableSkills}
-                  onInstall={handleInstallAvailableSkill}
-                  installing={installingSkill}
-                />
-              </section>
-            )}
-
-            {showLoadingState && (
-              <div className="flex flex-col items-center justify-center rounded-xl border border-dashed border-border/50 bg-muted/10 py-10 text-center">
-                <div className="mb-1 text-base font-medium">Loading...</div>
-                <div className="text-sm text-muted-foreground">Fetching skills catalog.</div>
-              </div>
-            )}
-
-            {!showLoadingState && skillCatalogError ? (
-              <div className="flex flex-col items-start gap-3 rounded-xl border border-destructive/40 bg-destructive/5 px-4 py-4 text-left">
-                <div className="flex items-center gap-2">
-                  <Badge variant="outline" className="border-destructive/40 text-destructive">
-                    Connection issue
-                  </Badge>
-                  <span className="text-sm text-destructive">{skillCatalogError}</span>
-                </div>
-                <Button size="sm" variant="outline" onClick={() => void refreshSkillsCatalog()}>
-                  Retry
-                </Button>
-              </div>
-            ) : null}
-
-            {!showLoadingState &&
-              !skillCatalogError &&
-              installations.length === 0 &&
-              availableSkills.length === 0 && (
-                <div className="flex flex-col items-center justify-center rounded-xl border border-dashed border-border/50 bg-muted/10 py-10 text-center">
-                  <div className="mb-1 text-base font-medium">No skills found</div>
-                  <div className="text-sm text-muted-foreground">
-                    {searchQuery
-                      ? "Try adjusting your search query."
-                      : "Install a skill to give Cowork superpowers."}
-                  </div>
-                </div>
-              )}
-          </div>
-        </InlineErrorBoundary>
-      </div>
-
-      <SkillDetailDialog workspaceId={workspaceId} />
-    </div>
-  );
-}
