--- apps/desktop/src/ui/skills/index.ts
@@ -1,1 +1,0 @@
-export { SkillsCatalogPage } from "./SkillsCatalogPage";
