+++ apps/desktop/test/bootstrap-cache.test.ts
@@ -477,7 +477,6 @@ function resetStoreToCachedSeed(value: unknown = cachedState) {
     tasksById: {},
     taskListLoadingByWorkspaceId: {},
     taskError: null,
-    pluginManagementWorkspaceId: null,
     workspaceRuntimeById: {},
     threadRuntimeById: {},
     latestTodosByThreadId: {},
@@ -638,8 +637,7 @@ describe("desktop bootstrap cache", () => {
     expect(seed?.bootstrapPending).toBe(true);
     expect(seed?.selectedWorkspaceId).toBe("ws-cached");
     expect(seed?.selectedThreadId).toBe("thread-cached");
-    expect(seed?.pluginManagementWorkspaceId).toBeNull();
-    expect(seed?.view).toBe("skills");
+    expect(seed?.view).toBe("settings");
     expect(seed?.sidebarCollapsed).toBe(true);
     expect(seed?.sidebarWidth).toBe(320);
     expect(seed?.contextSidebarCollapsed).toBe(true);
@@ -961,8 +959,7 @@ describe("desktop bootstrap cache", () => {
     const seed = buildCachedDesktopStateSeed(legacyCachedState);
     expect(seed?.selectedWorkspaceId).toBe("ws-cached");
     expect(seed?.selectedThreadId).toBe("thread-cached");
-    expect(seed?.pluginManagementWorkspaceId).toBeNull();
-    expect(seed?.view).toBe("skills");
+    expect(seed?.view).toBe("settings");
     expect(seed?.workspaces?.[0]?.wsProtocol).toBe("jsonrpc");
   });
 
@@ -996,32 +993,17 @@ describe("desktop bootstrap cache", () => {
     expect(seed?.desktopFeatureFlagOverrides?.REMOVEDUI).toBeUndefined();
   });
 
-  test("buildCachedDesktopStateSeed restores plugin management workspace selection", () => {
+  test("buildCachedDesktopStateSeed ignores legacy plugin management workspace fields", () => {
     const seed = buildCachedDesktopStateSeed({
       ...cachedState,
-      persistedState: {
-        ...cachedState.persistedState,
-        workspaces: [
-          ...cachedState.persistedState.workspaces,
-          {
-            id: "ws-management",
-            name: "Management Workspace",
-            path: "/tmp/workspace-management",
-            createdAt: "2026-03-19T00:00:00.000Z",
-            lastOpenedAt: "2026-03-19T00:00:00.000Z",
-            defaultEnableMcp: true,
-            defaultBackupsEnabled: true,
-            yolo: false,
-          },
-        ],
-      },
       ui: {
         ...cachedState.ui,
         pluginManagementWorkspaceId: "ws-management",
+        pluginManagementMode: "workspace",
       },
     });
 
-    expect(seed?.pluginManagementWorkspaceId).toBe("ws-management");
+    expect(seed).not.toHaveProperty("pluginManagementWorkspaceId");
     expect(seed?.selectedWorkspaceId).toBe("ws-cached");
   });
 
@@ -1142,8 +1124,7 @@ describe("desktop bootstrap cache", () => {
     expect(state.bootstrapPending).toBe(false);
     expect(state.selectedWorkspaceId).toBe("ws-live");
     expect(state.selectedThreadId).toBe("thread-live");
-    expect(state.pluginManagementWorkspaceId).toBeNull();
-    expect(state.view).toBe("skills");
+    expect(state.view).toBe("settings");
     expect(state.sidebarCollapsed).toBe(true);
   });
 
@@ -1321,94 +1302,6 @@ describe("desktop bootstrap cache", () => {
     expect(state.view).toBe("task");
     expect(state.selectedTaskId).toBeNull();
     expect(state.selectedThreadId).toBeNull();
-  });
-
-  test("init restores plugin management workspace selection from cached state", async () => {
-    const pluginManagementCachedState = {
-      ...cachedState,
-      persistedState: {
-        ...cachedState.persistedState,
-        workspaces: [
-          {
-            ...cachedState.persistedState.workspaces[0],
-            id: "ws-live",
-            name: "Live Workspace",
-            path: "/tmp/workspace-live",
-          },
-          {
-            id: "ws-live-management",
-            name: "Management Workspace",
-            path: "/tmp/workspace-live-management",
-            createdAt: "2026-03-20T00:00:00.000Z",
-            lastOpenedAt: "2026-03-20T00:00:00.000Z",
-            defaultEnableMcp: true,
-            defaultBackupsEnabled: true,
-            yolo: false,
-          },
-        ],
-        threads: [
-          {
-            ...cachedState.persistedState.threads[0],
-            id: "thread-live",
-            workspaceId: "ws-live",
-          },
-        ],
-      },
-      ui: {
-        ...cachedState.ui,
-        selectedWorkspaceId: "ws-live",
-        selectedThreadId: "thread-live",
-        pluginManagementWorkspaceId: "ws-live-management",
-      },
-    };
-    loadedState = {
-      ...loadedState,
-      workspaces: [
-        {
-          id: "ws-live",
-          name: "Live Workspace",
-          path: "/tmp/workspace-live",
-          createdAt: "2026-03-20T00:00:00.000Z",
-          lastOpenedAt: "2026-03-20T00:00:00.000Z",
-          defaultEnableMcp: true,
-          defaultBackupsEnabled: true,
-          yolo: false,
-        },
-        {
-          id: "ws-live-management",
-          name: "Management Workspace",
-          path: "/tmp/workspace-live-management",
-          createdAt: "2026-03-20T00:00:00.000Z",
-          lastOpenedAt: "2026-03-20T00:00:00.000Z",
-          defaultEnableMcp: true,
-          defaultBackupsEnabled: true,
-          yolo: false,
-        },
-      ],
-      threads: [
-        {
-          id: "thread-live",
-          workspaceId: "ws-live",
-          title: "Live Thread",
-          titleSource: "manual",
-          createdAt: "2026-03-20T00:00:00.000Z",
-          lastMessageAt: "2026-03-20T00:00:00.000Z",
-          status: "active",
-          sessionId: null,
-          messageCount: 0,
-          lastEventSeq: 0,
-        },
-      ],
-    };
-    localStorageMock.setItem(DESKTOP_STATE_CACHE_KEY, JSON.stringify(pluginManagementCachedState));
-    resetStoreToCachedSeed(pluginManagementCachedState);
-
-    await useAppStore.getState().init();
-
-    const state = useAppStore.getState();
-    expect(state.selectedWorkspaceId).toBe("ws-live");
-    expect(state.pluginManagementWorkspaceId).toBe("ws-live-management");
-    expect(state.view).toBe("skills");
   });
 
   test("init marks the selected startup thread as hydrating before deferred restore finishes", async () => {
