+++ apps/desktop/test/install-plugin-dialog.test.ts
@@ -75,8 +75,7 @@ const {
   InstallPluginDialog,
   shouldDisablePluginInstallForScope,
   shouldRequireFreshPluginPreviewForScope,
-} = await import("../src/ui/plugins/InstallPluginDialog");
-const { PluginsCatalogPage } = await import("../src/ui/plugins/PluginsCatalogPage");
+} = await import("../src/ui/settings/toolAccess/InstallPluginDialog");
 mock.restore();
 
 const workspaceId = "ws-plugins";
@@ -125,164 +124,7 @@ function baseWorkspaceState() {
   };
 }
 
-describe("plugins catalog page", () => {
-  test("shows loading state while plugins are loading", async () => {
-    const previousState = useAppStore.getState();
-    let root: ReturnType<typeof createRoot> | null = null;
-    useAppStore.setState({
-      ...baseWorkspaceState(),
-      workspaceRuntimeById: {
-        [workspaceId]: {
-          ...defaultWorkspaceRuntime(),
-          pluginsCatalog: null,
-          pluginsLoading: true,
-        },
-      },
-    } as any);
-
-    const harness = setupJsdom();
-    try {
-      (
-        harness.dom.window.HTMLElement.prototype as {
-          attachEvent?: () => void;
-          detachEvent?: () => void;
-        }
-      ).attachEvent = () => {};
-      (
-        harness.dom.window.HTMLElement.prototype as {
-          attachEvent?: () => void;
-          detachEvent?: () => void;
-        }
-      ).detachEvent = () => {};
-      const container = harness.dom.window.document.getElementById("root");
-      if (!container) throw new Error("missing root");
-      root = createRoot(container);
-
-      await act(async () => {
-        root.render(
-          createElement(PluginsCatalogPage, {
-            workspaceId,
-            searchQuery: "",
-            setSearchQuery: () => {},
-          }),
-        );
-      });
-
-      expect(container.textContent).toContain("Loading...");
-      expect(container.textContent).toContain("Fetching plugins catalog.");
-
-      await act(async () => {
-        root.unmount();
-      });
-    } finally {
-      if (root) {
-        await act(async () => {
-          root.unmount();
-        });
-      }
-      useAppStore.setState(previousState);
-      harness.restore();
-    }
-  });
-
-  test("shows inline error state when plugin catalog refresh fails", async () => {
-    const previousState = useAppStore.getState();
-    let root: ReturnType<typeof createRoot> | null = null;
-    useAppStore.setState({
-      ...baseWorkspaceState(),
-      workspaceRuntimeById: {
-        [workspaceId]: {
-          ...defaultWorkspaceRuntime(),
-          pluginsCatalog: null,
-          pluginsLoading: false,
-          pluginsError: "Unable to refresh plugins catalog.",
-        },
-      },
-    } as any);
-
-    const harness = setupJsdom();
-    try {
-      const container = harness.dom.window.document.getElementById("root");
-      if (!container) throw new Error("missing root");
-      root = createRoot(container);
-
-      await act(async () => {
-        root.render(
-          createElement(PluginsCatalogPage, {
-            workspaceId,
-            searchQuery: "",
-            setSearchQuery: () => {},
-          }),
-        );
-      });
-
-      expect(container.textContent).toContain("Connection issue");
-      expect(container.textContent).toContain("Unable to refresh plugins catalog.");
-      expect(container.textContent).toContain("Retry");
-
-      await act(async () => {
-        root.unmount();
-      });
-    } finally {
-      if (root) {
-        await act(async () => {
-          root.unmount();
-        });
-      }
-      useAppStore.setState(previousState);
-      harness.restore();
-    }
-  });
-
-  test("renders the new plugin install affordance", async () => {
-    const previousState = useAppStore.getState();
-    let root: ReturnType<typeof createRoot> | null = null;
-    useAppStore.setState({
-      ...baseWorkspaceState(),
-      workspaceRuntimeById: {
-        [workspaceId]: {
-          ...defaultWorkspaceRuntime(),
-          pluginsCatalog: {
-            warnings: [],
-            plugins: [],
-            availablePlugins: [],
-          },
-        },
-      },
-    } as any);
-
-    const harness = setupJsdom();
-    try {
-      const container = harness.dom.window.document.getElementById("root");
-      if (!container) throw new Error("missing root");
-      root = createRoot(container);
-
-      await act(async () => {
-        root.render(
-          createElement(PluginsCatalogPage, {
-            workspaceId,
-            searchQuery: "",
-            setSearchQuery: () => {},
-          }),
-        );
-      });
-
-      expect(container.textContent).toContain("+ New plugin");
-
-      await act(async () => {
-        root.unmount();
-      });
-    } finally {
-      if (root) {
-        await act(async () => {
-          root.unmount();
-        });
-      }
-      useAppStore.setState(previousState);
-      harness.restore();
-    }
-  });
-
+describe("install plugin dialog", () => {
   test("new plugin dialog opens from a clean state", async () => {
     const previousState = useAppStore.getState();
     let root: ReturnType<typeof createRoot> | null = null;
@@ -638,196 +480,5 @@ describe("plugins catalog page", () => {
         pluginInstallInFlight: false,
       }),
     ).toBe(true);
-  });
-
-  test("renders enabled and disabled plugin sections with counts", async () => {
-    const previousState = useAppStore.getState();
-    let root: ReturnType<typeof createRoot> | null = null;
-    useAppStore.setState({
-      ...baseWorkspaceState(),
-      workspaceRuntimeById: {
-        [workspaceId]: {
-          ...defaultWorkspaceRuntime(),
-          pluginsCatalog: {
-            warnings: [],
-            plugins: [
-              {
-                id: "workspace-tools",
-                name: "workspace-tools",
-                displayName: "Workspace Tools",
-                description: "Workspace artifact helpers",
-                scope: "workspace",
-                discoveryKind: "marketplace",
-                installed: true,
-                enabled: true,
-                rootDir: "/tmp/plugin-workspace/.agents/plugins/workspace-tools",
-                manifestPath:
-                  "/tmp/plugin-workspace/.agents/plugins/workspace-tools/.codex-plugin/plugin.json",
-                skillsPath: "/tmp/plugin-workspace/.agents/plugins/workspace-tools/skills",
-                skills: [
-                  {
-                    name: "workspace-tools:documents",
-                    description: "Create documents",
-                    enabled: true,
-                  },
-                  {
-                    name: "workspace-tools:presentations",
-                    description: "Create presentations",
-                    enabled: true,
-                  },
-                  {
-                    name: "workspace-tools:spreadsheets",
-                    description: "Create spreadsheets",
-                    enabled: true,
-                  },
-                ],
-                mcpServers: [],
-                apps: [{ id: "figma-app", displayName: "Figma App", description: "Metadata only" }],
-                warnings: [],
-              },
-              {
-                id: "plugin-2",
-                name: "slack-toolkit",
-                displayName: "Slack Toolkit",
-                description: "Slack helpers",
-                scope: "user",
-                discoveryKind: "direct",
-                installed: true,
-                enabled: false,
-                rootDir: "/tmp/home/.agents/plugins/slack-toolkit",
-                manifestPath: "/tmp/home/.agents/plugins/slack-toolkit/.codex-plugin/plugin.json",
-                skillsPath: "/tmp/home/.agents/plugins/slack-toolkit/skills",
-                skills: [],
-                mcpServers: ["slack"],
-                apps: [],
-                warnings: [],
-              },
-            ],
-            availablePlugins: [],
-          },
-        },
-      },
-    } as any);
-
-    const harness = setupJsdom();
-    try {
-      const container = harness.dom.window.document.getElementById("root");
-      if (!container) throw new Error("missing root");
-      root = createRoot(container);
-
-      await act(async () => {
-        root.render(
-          createElement(PluginsCatalogPage, {
-            workspaceId,
-            searchQuery: "",
-            setSearchQuery: () => {},
-          }),
-        );
-      });
-
-      expect(container.textContent).toContain("Enabled");
-      expect(container.textContent).toContain("Disabled");
-      expect(container.textContent).toContain("Workspace Tools");
-      expect(container.textContent).toContain("Slack Toolkit");
-      expect(container.textContent).toContain("3 skills");
-      expect(container.textContent).not.toContain("1 apps");
-
-      await act(async () => {
-        root.unmount();
-      });
-    } finally {
-      if (root) {
-        await act(async () => {
-          root.unmount();
-        });
-      }
-      useAppStore.setState(previousState);
-      harness.restore();
-    }
-  });
-
-  test("renders duplicate plugin ids from different scopes without collapsing a card", async () => {
-    const previousState = useAppStore.getState();
-    let root: ReturnType<typeof createRoot> | null = null;
-    useAppStore.setState({
-      ...baseWorkspaceState(),
-      workspaceRuntimeById: {
-        [workspaceId]: {
-          ...defaultWorkspaceRuntime(),
-          pluginsCatalog: {
-            warnings: [],
-            plugins: [
-              {
-                id: "figma-toolkit",
-                name: "figma-toolkit",
-                displayName: "Workspace Figma Toolkit",
-                description: "Workspace helpers",
-                scope: "workspace",
-                discoveryKind: "direct",
-                installed: true,
-                enabled: true,
-                rootDir: "/tmp/plugin-workspace/.agents/plugins/figma-toolkit",
-                manifestPath:
-                  "/tmp/plugin-workspace/.agents/plugins/figma-toolkit/.codex-plugin/plugin.json",
-                skillsPath: "/tmp/plugin-workspace/.agents/plugins/figma-toolkit/skills",
-                skills: [],
-                mcpServers: [],
-                apps: [],
-                warnings: [],
-              },
-              {
-                id: "figma-toolkit",
-                name: "figma-toolkit",
-                displayName: "User Figma Toolkit",
-                description: "Global helpers",
-                scope: "user",
-                discoveryKind: "direct",
-                installed: true,
-                enabled: false,
-                rootDir: "/tmp/home/.agents/plugins/figma-toolkit",
-                manifestPath: "/tmp/home/.agents/plugins/figma-toolkit/.codex-plugin/plugin.json",
-                skillsPath: "/tmp/home/.agents/plugins/figma-toolkit/skills",
-                skills: [],
-                mcpServers: [],
-                apps: [],
-                warnings: [],
-              },
-            ],
-            availablePlugins: [],
-          },
-        },
-      },
-    } as any);
-
-    const harness = setupJsdom();
-    try {
-      const container = harness.dom.window.document.getElementById("root");
-      if (!container) throw new Error("missing root");
-      root = createRoot(container);
-
-      await act(async () => {
-        root.render(
-          createElement(PluginsCatalogPage, {
-            workspaceId,
-            searchQuery: "",
-            setSearchQuery: () => {},
-          }),
-        );
-      });
-
-      const pageText = container.textContent ?? "";
-      expect(pageText).toContain("Workspace Figma Toolkit");
-      expect(pageText).toContain("User Figma Toolkit");
-      expect(pageText).toContain("Workspace plugin");
-      expect(pageText).toContain("Library plugin");
-    } finally {
-      if (root) {
-        await act(async () => {
-          root.unmount();
-        });
-      }
-      useAppStore.setState(previousState);
-      harness.restore();
-    }
   });
 });
