+++ apps/desktop/test/workspace-settings-sync.harness.ts
@@ -648,7 +648,6 @@ export function registerWorkspaceSettingsSyncLifecycleHooks() {
       ],
       threads: [],
       selectedWorkspaceId: workspaceId,
-      pluginManagementWorkspaceId: null,
       selectedThreadId: null,
       workspaceRuntimeById: {},
       threadRuntimeById: {},
